feat(crux_macros): be explicit in effect macro about typegen kind
