+++ crux_macros/src/effect/mod.rs
@@ -1,2 +1,3 @@
-pub mod facet;
-pub mod serde;
+pub mod macro_impl;
+#[cfg(test)]
+pub mod tests;
