--- crux_macros/src/effect/serde/mod.rs
@@ -1,3 +1,0 @@
-pub mod macro_impl;
-#[cfg(test)]
-pub mod tests;
