+++ examples/bridge_echo/shared/Cargo.toml
@@ -13,11 +13,11 @@ typegen = ["crux_core/typegen"]
 [dependencies]
 crux_core.workspace = true
 serde = { workspace = true, features = ["derive"] }
-uniffi = "0.29.2"
+uniffi = "0.29.3"
 wasm-bindgen = "0.2.100"
 
 [target.uniffi-bindgen.dependencies]
-uniffi = { version = "0.29.2", features = ["cli"] }
+uniffi = { version = "0.29.3", features = ["cli"] }
 
 [build-dependencies]
-uniffi = { version = "0.29.2", features = ["build"] }
+uniffi = { version = "0.29.3", features = ["build"] }
