--- examples/cat_facts/Android/.settings/org.eclipse.buildship.core.prefs
@@ -1,13 +1,0 @@
-arguments=--init-script /var/folders/vj/_19mcln57qz5j75lqqjtxc540000gn/T/db3b08fc4a9ef609cb16b96b200fa13e563f396e9bb1ed0905fdab7bc3bc513b.gradle
-auto.sync=false
-build.scans.enabled=false
-connection.gradle.distribution=GRADLE_DISTRIBUTION(WRAPPER)
-connection.project.dir=
-eclipse.preferences.version=1
-gradle.user.home=
-java.home=/Users/stuartharris/Library/Java/JavaVirtualMachines/openjdk-20.0.2/Contents/Home
-jvm.arguments=
-offline.mode=false
-override.workspace.settings=true
-show.console.view=true
-show.executions.view=true
